+++ src/app/_components/common/daily/DailyUserTitle.tsx
@@ -7,7 +7,7 @@ import Modal from '@/app/_components/common/modal/Modal';
 import KebabMenuIcon from '@/app/_components/icon/kebabMenu/KebabMenuIcon';
 import { MODAL_TYPE_MAP } from '@/app/_constants/modal';
 import { getCookies } from '@/app/_store/cookie/cookies';
-import { usePostDeletedStore } from '@/app/_store/isPostDeleted';
+import { usePostDeletedStore } from '@/app/_store/isPostDeletedStore';
 import { useRouter } from 'next/navigation';
 import { useState } from 'react';
