+++ src/app/_store/isImageDeletedStore.ts
